Run receiver and generator in parallel
The receiver requests the files it wants, but for any that the sender
can't open, it simply doesn't reply. So, sometimes after sending a file
list request, there's simply no response, and no way to know whether it
just hasn't arrived yet, or isn't coming.

This could be handled by wiring through an indication there's no input
yet, but it seems better to now move to running the sender and generator
in parallel, analogous to in rsync. In rsyn they're not forked processes
but rather separate threads.

This code successfully reads (and discards) the contents of remote trees
including where some files can't be read.