Experiment with using `tracing`
It looks promising, but also seems a bit less mature than log/fern, so I
won't convert for now. In particular it seems you can't have different
sinks with different levels, and it's a bit confusing how to make it log
to a file. I'm concerned there might be more of a performance hit, too.

Let's leave it until it's more mature or more obviously needed.

Colored output by default is nice, though.